Restructure files to be more organized
