+++ .gitmodules
@@ -1,5 +1,5 @@
 [submodule "JUCE"]
-	path = JUCE
+	path = lib/JUCE
 	url = https://github.com/juce-framework/JUCE.git
 [submodule "mosquitto"]
 	path = mosquitto
@@ -8,8 +8,8 @@
 	path = openssl
 	url = git://git.openssl.org/openssl.git
 [submodule "RS-232"]
-	path = RS-232
+	path = lib/RS-232
 	url = https://gitlab.com/Teuniz/RS-232.git
 [submodule "fmt"]
-	path = fmt
+	path = lib/fmt
 	url = git@github.com:fmtlib/fmt.git
