--- Profiler/CMakeLists.txt
@@ -1,6 +1,0 @@
-cmake_minimum_required(VERSION 3.15)
-
-target_sources(SuperMileageDisplay PRIVATE
-	Profiler.h
-	Profiler.cpp
-)
