--- Serial/CMakeLists.txt
@@ -1,7 +1,0 @@
-cmake_minimum_required(VERSION 3.15)
-
-target_compile_features(SuperMileageDisplay PRIVATE cxx_std_20)
-target_sources(SuperMileageDisplay PRIVATE 
-	SerialClient.h
-	SerialClient.cpp
-)
