+++ include/Display/MapComponent.h
