+++ src/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectory(Display)
+add_subdirectory(Profiler)
+#add_subdirectory(Serial)
