+++ src/Display/CMakeLists.txt
@@ -1,0 +1,14 @@
+target_sources(
+    SuperMileageDisplay
+    PRIVATE
+    CMakeLists.txt
+    ColorLight.cpp
+    LapCounter.cpp
+    MainComponent.cpp
+    Main.cpp
+    MapComponent.cpp
+    RaceTimer.cpp
+    Speedometer.cpp
+    Tiltmeter.cpp
+    VerticleGauge.cpp
+)
