+++ src/Display/LapCounter.cpp
@@ -1,9 +1,8 @@
-#include "LapCounter.h"
-#include <sstream>
-#include <locale>
+#include "Display/LapCounter.h"
+#include <fmt/format.h>
+#include <string>
 
-#include "Profiler.h"
-#include "Formatter.h"
+#include "Profiler/Profiler.h"
 
 LapCounter::LapCounter(double lapLength, unsigned lapAmount) :
 	_lapCounter(_lapCount), _lapProgress(_lapDist), _lapCount(0), 
@@ -52,10 +51,8 @@ void LapCounter::incDistanceTraveled(double dist)
 	if (!_finished)
 	{
 		_lapDist += dist / _lapLength;
-		std::stringstream ss;
-		ss.imbue(std::locale(ss.getloc(), new Formatter));
-		ss << "Distance: " << (int)((_lapDist * _lapLength) * 5280.0) << " ft";
-		_lapProgress.setTextToDisplay(ss.str());
+		std::string label = fmt::format("%d ft", _lapDist);
+		_lapProgress.setTextToDisplay(label);
 
 		if (_lapDist >= 1.0)
 		{
