+++ src/Profiler/CMakeLists.txt
@@ -1,0 +1,1 @@
+target_sources(SuperMileageDisplay PRIVATE Profiler.cpp)
