+++ src/Serial/CMakeLists.txt
@@ -1,0 +1,1 @@
+target_sources(SerialClient.cpp)
